+++ packages/components/pages/f-contact-preferences/src/components/_tests/ContactPreferences.test.js
@@ -76,6 +76,7 @@ describe('ContactPreferences Component', () => {
         };
         sutProps = {
             authToken: token,
+            isAuthFinished: true,
             smartGatewayBaseUrl: baseUrl
         };
     });