+++ packages/components/pages/f-contact-preferences/CHANGELOG.md
@@ -4,12 +4,25 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.8.0
+------------------------------
+*November 25, 2021*
+
+### Added
+- Expose new isAuthFinished `required` property
+
+### Changed
+- Registered the Store Module on Create()
+- Initialise Page based on isAuthFinished
+
+
 v0.7.0
 ------------------------------
 *November 23, 2021*
 
 ### Changed
 - Rename boolean fields to match naming conventions
+
 
 ### Added
 - Added Accessibility Tests.